HW10: move Event to domain/ folder
